+++ changelog.md
@@ -5,6 +5,7 @@ Codion Change Log
 ### is.codion.swing.common.ui
 - AbstractButtonBuilder now sets the initial background, foreground and font specified by the action.
 - AbstractControl now overrides putValue() and getValue() in order to prevent modification of the 'enabled' property.
+- Control.Builder.value() added.
 
 ## 0.17.21
 ### is.codion.swing.common.model